email_parser/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
